Add Luna as a page
Add the the basic UI and audio for Luna as a page. The big change here is that
the initial patch is loaded when a user visits the Luna page. This means moving
the initial patch out of TypeScript into Elm.
Modularize audio implementation to allow us to load other instruments down the
road.
Rename image directory and update asset paths.
Install Elm JSON packages and parcel static files plugin.
Controllers and tuning still need to be added.