Add patch loading from patch browser
When a user selects a patch from the patch browser, we call over a port to
attempt to retrieve it. For now, this is just retrieved from the in-memory
patch list loaded over HTTP at init.
Add encoders for patch metadata to make the request.
Handle load request on JavaScript side. Follow a patch load with an instrument
patch, which is what configures the web audio instrument.
Modify the placeholder patches for testing purposes.