Add auth and patch storage
Add web native for auth and patch storage.
Add Session to to represent auth state and wrap a user creator.
Add login with Fission to replace editor icons when a user is not signed in.
Show creator selector only when user is logged in.
Add patch persistence for user patches. A user can store and load patches to
web native storage. Their stored patches are loaded immediately after the are
authenticated on page load or refresh.
Move GotPatches to Shared. All user patches will be collected up on app init
for now. We may need to optimize this some day.
Add a ligther purple to get better contrast on controls where we have a purple
background.