Ignore stop notes if note is inactive
We may get stop note calls from a computer keyboard controller if a user
presses a modifier key, presses a playable key while holding the modifier key,
releases the modifier key, and finally releases the playable key. We might be
able to handle this in the controller code, but this option is simpler and
works for now.